MA-7399 - [Android_MX6QP_SD]Video:Some streams played unfluently if
connect to audio receiver, 100%

The issue occues when passthrough is not checked in AudioRoute.apk

44100 pcm sent to hdmi, but comsumed by 48000.
The video will jump frames to catch up audio.

The framework will auto check direct output profile.
If has, then use AUDIO_OUTPUT_FLAG_DIRECT.

So when passthrough is not checked, the OMX will decode ac3 to 44.1k pcm.
But the audiopolicy use direct output as 48K.

In audiopolicy, when check direct output,
if hdmi, also assure that the application enable direct output.

Signed-off-by: fang hui <b31070@freescale.com>